requirements.txt
Flask==3.0.3
openai==1.14.3
PyPDF2
python-docx
werkzeug
python-dotenv